Разделила интерфейсы
