+++ src/model/Datatable.java
@@ -2,12 +2,7 @@ package model;
 
 import java.io.IOException;
 
-public interface Serviceble {
-    void addHuman(String family, String name);
-   // public getByHuman ();
-    void sortByName();
-    void sortByFamily();
-
+public interface Datatable {
     public void save(Object object) throws IOException;
     public Object read() throws ClassNotFoundException, IOException;
 //    void save(Writable writable) throws IOException;
