+++ src/model/Service.java
@@ -1,0 +1,12 @@
+package model;
+
+import java.io.IOException;
+
+public interface Service {
+    void addHuman(String family, String name);
+   // public getByHuman ();
+    void sortByName();
+    void sortByFamily();
+
+
+}
